requirements.txt
numpy==1.18.5
tensorflow-gpu==2.3.0
pydot
Pillow==8.0.1

# Application wise
matplotlib==3.3.2
opencv-python==4.4.0.44
gdown==3.13.0
scikit-image